% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_cat}
\alias{annotate_cat}
\title{Generate colors and ids for categorical annotations}
\usage{
annotate_cat(df, col = NULL, base = NULL, sort_label = T,
  na_val = "ZZ_Missing", colorset = "rainbow", color_order = "sort")
}
\arguments{
\item{df}{data frame to annotate}

\item{col}{name of the character column to annotate}

\item{base}{base name for the annotation, which wil be used in the desc table. If not provided, will use col as base.}

\item{sort_label}{a logical value to determine if the data in col should be arranged alphanumerically before ids are assigned. default = T.}

\item{na_val}{The value to use to replace NAs. default = "ZZ_Missing".}

\item{colorset}{The colorset to use for assigning category colors. Options are "rainbow","viridis","inferno","magma", and "terrain"}

\item{color_order}{The order in which colors should be assigned. Options are "sort" and "random". "sort" assigns colors in order; "random" will randomly assign colors.}
}
\value{
A modified data frame: the annotated column will be renamed base_label, and base_id and base_color columns will be appended
}
\description{
Generate colors and ids for categorical annotations
}
\examples{
flowers <- iris \%>\%
  annotate_cat(Species)

head(flowers)

flowers2 <- iris \%>\%
 annotate_num(Species, spp, sort_label = F, colorset = "viridis")
 
head(flowers2)
}
