% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_num}
\alias{annotate_num}
\title{Generate colors and ids for numeric annotations}
\usage{
annotate_num(df, col = NULL, base = NULL, scale = "log10", na_val = 0,
  colorset = c("darkblue", "white", "red"))
}
\arguments{
\item{df}{data frame to annotate}

\item{col}{name of the numeric column to annotate}

\item{base}{base name for the annotation, which wil be used in the desc table. If not provided, will use col as base.}

\item{scale}{The scale to use for assigning colors. Options are "linear","log10","log2, and "zscore"}

\item{na_val}{The value to use to replace NAs. default = 0.}

\item{colorset}{A vector of colors to use for the color gradient. default = c("darkblue","white","red")}
}
\value{
A modified data frame: the annotated column will be renamed base_label, and base_id and base_color columns will be appended
}
\description{
Generate colors and ids for numeric annotations
}
\examples{
cars <- mtcars \%>\%
  annotate_num(wt)

head(cars)

cars2 <- mtcars \%>\%
 annotate_num(wt, weight, "linear", colorset = c("#000000","#808080","#FF0000"))
 
head(cars2)
}
