% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getucscdata.R
\name{bed_to_GRanges}
\alias{bed_to_GRanges}
\title{Convert data.frames in BED-like format to GRanges objects}
\usage{
bed_to_GRanges(bed)
}
\description{
Convert data.frames in BED-like format to GRanges objects
}

