% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{build_feather}
\alias{build_feather}
\title{Write scrattch data to feather files}
\usage{
build_feather(anno = NULL, data = NULL, desc = NULL, feather_dir = NULL)
}
\description{
Write scrattch data to feather files
}
