% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{build_header_labels}
\alias{build_header_labels}
\title{Build colorful, rectangular labels for plot headers in plot space}
\usage{
build_header_labels(data, ngenes, nsamples, nclust, labelheight = 25,
  labeltype = "simple")
}
\description{
Build colorful, rectangular labels for plot headers in plot space
}
