% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{build_header_polygons}
\alias{build_header_polygons}
\title{Build polygons from plot data for fancy headers built into the plot area}
\usage{
build_header_polygons(data, ngenes, nsamples, nclust, labelheight = 25,
  labeltype = "angle")
}
\description{
Build polygons from plot data for fancy headers built into the plot area
}
