% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{build_palette}
\alias{build_palette}
\title{Generate a palette of related colors around a central color}
\usage{
build_palette(central_color)
}
\arguments{
\item{central_color}{a single color value around which to build the palette}
}
\value{
a list of palette results, containing:
\itemize{
\item palette: a character vector with a 10-color palette realted to the central_color.
\item palette_plot: a ggplot2 object showing the 10 colors in palette.
\item colorset: A character vector with the full set of 100 colors from which the palette was selected.
\item colorset_plot: A ggplot2 object displaying all 100 colors in the colorset.
\item weave: a ggplot2 object generated with weave_plot() showing the intersection of all of the collors in the palette.
}
}
\description{
Generate a palette of related colors around a central color
}
