% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/databasehelp.R
\name{check_db_structure}
\alias{check_db_structure}
\title{Check Database structure}
\usage{
check_db_structure(db.file, verbose = T)
}
\arguments{
\item{db.file}{a character object that gives the file location of a SQLite3 database.}

\item{verbose}{a logical value indicating if errors should be displayed. If set to FALSE, all messages will be suppressed.}
}
\value{
A logical value indicating if the database passes all required checks.
}
\description{
This function will examine a SQLite3 database to make sure it's consistent
with the structure required for my Shiny applications.
}
\examples{
database_file <- "//AIBSData2/mct-t200/ShinyApps/linnarsson/linnarsson2.db"

check_db_structure(database_file)

check_db_structure(database_file,verbose=F)
}

