% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{check_feather_structure}
\alias{check_feather_structure}
\title{Check Feather Directory structure}
\usage{
check_feather_structure(feather_dir, verbose = T)
}
\arguments{
\item{feather_dir}{a character object that gives the file location of a feather dataset.}

\item{verbose}{a logical value indicating if errors should be displayed. If set to FALSE, all messages will be suppressed.}
}
\value{
A logical value indicating if the database passes all required checks.
}
\description{
This function will examine a SQLite3 database to make sure it's consistent
with the structure required for my Shiny applications.
}
\examples{
database_file <- "//AIBSData2/mct-t200/ShinyApps/linnarsson/linnarsson2.db"

check_db_structure(feather_dir)

check_db_structure(feather_dir,verbose=F)
}
