% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{check_list_structure}
\alias{check_list_structure}
\title{Check List structure}
\usage{
check_list_structure(list_data, verbose = T)
}
\arguments{
\item{list_data}{- a list object that contains data, anno, and desc tables.}

\item{verbose}{a logical value indicating if errors should be displayed. If set to FALSE, all messages will be suppressed.}
}
\value{
A logical value indicating if the list data passes all required checks.
}
\description{
This function will examine a list of tables to make sure it's consistent
with the structure required for scrattch applications.
}
\examples{
dataset_list <- list(data = v1_data, anno = v1_anno, desc = v1_desc)

check_list_structure(dataset_list)

check_db_structure(dataset_list,verbose=F)
}
