% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{chr_to_num}
\alias{chr_to_num}
\title{Evaluate a character string to a numeric vector}
\usage{
chr_to_num(in_chr)
}
\arguments{
\item{in_chr}{a character string}
}
\value{
a numeric object
}
\description{
Evaluate a character string to a numeric vector
}
\examples{
test <- "1:5"
chr_to_num(test)

test <- "8,7,45,20"
chr_to_num(test)
}
