% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plothelp.R
\name{chr_to_sql}
\alias{chr_to_sql}
\title{Convert a character object to a format compatible with SQL SELECT}
\usage{
chr_to_sql(in_chr)
}
\arguments{
\item{in_chr}{a character vector}
}
\value{
a character vector consisting of a single string with each object separated by
escaped quotes, bracketted by parens.
}
\description{
Convert a character object to a format compatible with SQL SELECT
}
\examples{
test <- c("Hspa8","Cux2","Snap25","Rspo2")
chr_to_sql(test)

sql_query <- paste0("SELECT * FROM my_table WHERE gene IN ",chr_to_sql(test))
print(sql_query)
}

