% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_sum}
\alias{color_sum}
\title{Mix two colors additively in RGB space}
\usage{
color_sum(col1, col2)
}
\arguments{
\item{col1}{A hex or R color}

\item{col2}{Another hex or R color}
}
\value{
The sum of col1 and col2 as a character hex color (e.g. "#FFFFFF")
}
\description{
Mix two colors additively in RGB space
}
\examples{
color_sum("red","green")

color_sum("#1B9E77","#D95F02")
}
