% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{db_to_list}
\alias{db_to_list}
\title{Read a SQLite database into a list}
\usage{
db_to_list(db.file, get_tables = NULL, genes = NULL, group_by = NULL,
  groups = NULL)
}
\arguments{
\item{db.file}{- a character object specifying the location of the database}

\item{genes}{- a character object with a list of genes to retrieve. NULL retrieves all genes. default = NULL}

\item{group_by}{- a character object with the base of the annotation column to filter on if only specific groups should be selected. NULL retrieves all groups. default = NULL}

\item{groups}{- a numeric object with the group indexes to be selected. NULL retrieves all groups. default = NULL}
}
\value{
A list object containing data.frames for each table in the database, filtered according to genes, group_by, and groups if provided.
}
\description{
This function will read a scrattch SQLite database into memory as a list.
}
\details{
Each table (desc, anno, and data) will be a data frame within the list.
}
