% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plothelp.R
\name{fix_mouse_genes}
\alias{fix_mouse_genes}
\title{Correct case of mouse gene symbols}
\usage{
fix_mouse_genes(genes)
}
\arguments{
\item{genes}{a character vector containing gene symbols}
}
\value{
a character vector containing genes with proper case
}
\description{
Correct case of mouse gene symbols
}
\examples{
test <- c("a330070k13rik","HDAC1","sncg","SeRPinB11")
fix_mouse_genes(test)
}

