% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{get_db_data}
\alias{get_db_data}
\title{Get data from a SQLite3 database for scrattch plotting}
\usage{
get_db_data(data_source, genes, group_by, clusters)
}
\description{
Get data from a SQLite3 database for scrattch plotting
}
