% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{get_feather_data}
\alias{get_feather_data}
\title{Get data directly from a directory of feather files}
\usage{
get_feather_data(feather_dir, genes, group_by, clusters)
}
\description{
Get data directly from a directory of feather files
}
