% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{get_internal_data}
\alias{get_internal_data}
\title{Combine data for plotting from the included Mouse V1 data}
\usage{
get_internal_data(genes, grouping, clusters)
}
\description{
Combine data for plotting from the included Mouse V1 data
}

