% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degenes.R
\name{get_sparse_pairwise_deg}
\alias{get_sparse_pairwise_deg}
\title{Find sparse, differentially-expressed genes in pairwise comparisons}
\usage{
get_sparse_pairwise_deg(data = v1_data, anno = v1_anno,
  group_by = "final", class1_groups = 1:23, class2_groups = 24:42,
  data_cut_up = 0, data_cut_dn = 0, frac_cut_up = 0.95,
  frac_cut_dn = 0.05, top = NULL, give = "genes")
}
\arguments{
\item{data}{- a data.frame containing the data to be filtered. First column must be "gene", followed by one column per sample.}

\item{anno}{- a data.frame with annotations for each sample.}

\item{group_by}{- a character object with the base of the annotation to be used for selecting groups.}

\item{class1_groups}{- a numeric object with the ids of the groups to be used as class 1}

\item{class2_groups}{- a numeric object with the ids of the groups to be used as class 2}

\item{data_cut_up}{- the minimum value cutoff for the "high" expression group. default: 0}

\item{data_cut_dn}{- the maximum value cutoff for the "low" expression group. default: 0}

\item{frac_cut_up}{- the minimum fraction of samples in the "high" expression group above data_cut_up. default: 0.95}

\item{frac_cut_dn}{- the maximum fraction of samples in the "low" expression group above data_cut_dn. default: 0.05}

\item{top}{- the number of results to keep from each class. default: NULL (returns all results)}

\item{give}{- either "table" or "genes".}
}
\value{
Depends on the give parameter: if "table", a data.frame with details for each gene.
if "genes", a character vector of the genes that pass the filters.
}
\description{
This function compares two groups or sets of groups to find genes that are differentially expressed based on
the fraction of cells with expression above a given thresshold. Because multiple groups can be combined, these
are termed "class 1" and "class 2" for this comparison. The comparison is performed in both directions, so will
identify genes with preferrential expression for both class 1 and class 2.
}
