% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getucscdata.R
\name{get_tss_regions}
\alias{get_tss_regions}
\title{Retrieve regions around TSS from the UCSC Genome Browser}
\usage{
get_tss_regions(symbols, expand = 5000, genome = "mm10")
}
\arguments{
\item{symbols}{A character vector containing gene symbols}

\item{expand}{A numeric vector with how many bp to expand upstream and downstream of the TSS. If a single number is supplied, it will be applied both 5' and 3' of the TSS. If 2 numbers are used, the first will be used for 5' and the second for 3'.}

\item{genome}{A character object with the genome abbreviation e.g. "mm10", "hg19".}
}
\value{
A data frame with BED-format columns chr, start, end, name, score, and strand.
}
\description{
Retrieve regions around TSS from the UCSC Genome Browser
}
\examples{
my_genes <- c("Hspa8","Sst","Ntsr1","Foxp2")
regions_5k <- get_tss_regions(my_genes)

regions_1kup_5kdn <- get_tss_regions(my_genes,
                                    expand=c(1000,5000),
                                    genome="mm9")
}

