% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{group_annotations}
\alias{group_annotations}
\title{Group annotation columns}
\usage{
group_annotations(df, keep_order = TRUE)
}
\arguments{
\item{df}{the annotation dataframe to arrange}

\item{keep_order}{a logical value. If FALSE, will sort the annotations alphanumerically by base.}
}
\value{
an annotation data frame with reordered columns
}
\description{
Group annotation columns
}
\examples{
anno <- mtcars \%>\%
 mutate(sample_id = paste0("car",1:n())) \%>\%
 select(sample_id, wt, mpg) \%>\%
 annotate_num(wt) \%>\%
 annotate_num(mpg)
 
head(anno)

anno2 <- group_annotations(anno)

head(anno2)

}
