% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{group_river_plot}
\alias{group_river_plot}
\title{Build Sankey Plots for up to four annotations}
\usage{
group_river_plot(data_source, group1 = NULL, group1_filter = NULL,
  group2 = NULL, group2_filter = NULL, group3 = NULL,
  group3_filter = NULL, group4 = NULL, group4_filter = NULL,
  width = 1000, height = 500)
}
\arguments{
\item{width}{numeric, plot width in pixels.}

\item{height}{numeric, plot height in pixels}
}
\value{
a NetworkD3 Sankey plot.
}
\description{
Builds a Sankey/River plot using NetworkD3 based on data annotations.
}
