% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{hclust_to_seg}
\alias{hclust_to_seg}
\title{Covert hclust objects to segments for use in ggplots}
\usage{
hclust_to_seg(hc, tree.dir = "down", dir.lims = c(0, 1))
}
\arguments{
\item{hc}{a hclust object}

\item{tree.dir}{a character object with the direction the tree points to, from root to leaves. options are "down" (default), "up","left", "right".}

\item{dir.lims}{a 2-member vector with the space in the direction of plotting that the dendrogram will occupy. default = c(0,1)}
}
\value{
a data.frame with segment values for ggplot2's geom_seg. columns: "x","xend","y","yend".
}
\description{
Covert hclust objects to segments for use in ggplots
}
