% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildplots.R
\name{heatcell_plot}
\alias{heatcell_plot}
\title{Heatmaps of gene expression of individual cells}
\usage{
heatcell_plot(genes = c("Hspa8", "Snap25", "Gad2", "Slc17a6"),
  clusters = 1:49, data_source = "internal", sort = F, logscale = T,
  normalize_rows = F, fontsize = 7, labelheight = 25)
}
\arguments{
\item{genes}{A character vector containing gene symbols to be plotted}

\item{clusters}{A numeric vector containing clusters to plot (for v1_anno, the range is 1:49)}

\item{data_source}{A character object defining where the data is stored. Currently only works with "internal"}

\item{logscale}{Logical object, determines if data is log scaled before plotting.}

\item{normalize_rows}{Logical object, determines if data are normalized to the maximum value for each gene. If FALSE, the heatmap is normalized to the maximum value across all genes.}

\item{fontsize}{numeric object, the font size (in pts) used to make the plot.}

\item{labelheight}{numeric object, Percent of the plot height that should be used for the labels (0 to 100).}
}
\value{
a ggplot2 plot object
}
\description{
This function will generate plots similar to those in the shiny heatmap geneterator.
Warning: this is currently only able to work with internally-supplied datasets (v1_data and v1_anno).
Extension to user-supplied datasets will come soon.
}
\examples{
heatcell_plot()

my_genes <- c("Ercc6","Ercc8","Trp53","Pgbd5")
my_clusters <- c(1,5,9,10,24,37)
heatcell_plot <- manybar_plot(my_genes,my_clusters,norm=T,font=12)

ggsave("plot_output.pdf",my_manybar_plot,height=0.2*length(my_genes)+2,width=4)
}

