% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildplots.R
\name{heater_plot}
\alias{heater_plot}
\title{Heatmaps of gene expression for clusters}
\usage{
heater_plot(genes = c("Hspa8", "Snap25", "Gad2", "Slc17a6"),
  clusters = 1:49, grouping = "final", calculation = "mean",
  data_source = "internal", normalize_rows = FALSE, logscale = T,
  fontsize = 7, toplabel = T, toptext = T, labelheight = 25,
  maxlabel = T, maxval = "auto", colorset = c("darkblue", "dodgerblue",
  "gray80", "orange", "orangered"))
}
\arguments{
\item{genes}{A character vector containing gene symbols to be plotted}

\item{clusters}{A numeric vector containing clusters to plot (for v1_anno, the range is 1:49)}

\item{calculation}{A character object stating which calculation to perform for each heatmap cell. Options: "mean" and "trimmed_mean"}

\item{data_source}{A character object defining where the data is stored. Currently only works with "internal"}

\item{normalize_rows}{Logical object, determines if data are normalized to the maximum value for each gene. If FALSE, the heatmap is normalized to the maximum value across all genes.}

\item{logscale}{Logical object, determines if data is log scaled before plotting.}

\item{fontsize}{numeric object, the font size (in pts) used to make the plot.}

\item{labelheight}{numeric object, Percent of the plot height that should be used for the labels (0 to 100).}
}
\value{
a ggplot2 plot object
}
\description{
Generates heatmaps for each gene and each cluster based on mean gene expression (more functions will be added later).
Warning: this is currently only able to work with internally-supplied datasets (v1_data and v1_anno).
Extension to user-supplied datasets will come soon.
}
\examples{
heater_plot()

my_genes <- c("Ercc6","Ercc8","Trp53","Pgbd5")
my_clusters <- c(1,5,9,10,24,37)
heater_plot(my_genes,my_clusters,logscale=T,fontsize=14)
}

