% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{heatmap_legend_plot}
\alias{heatmap_legend_plot}
\title{Build a heatmap legend plot}
\usage{
heatmap_legend_plot(minval = 0, maxval = 4, scale_name = "FPKM",
  colorset = c("darkblue", "dodgerblue", "gray80", "orange", "orangered"))
}
\arguments{
\item{minval}{numeric, the minimum value in the plot scale (default = 0)}

\item{maxval}{numeric, the maximum value in the plot scale (default = 4)}

\item{scale_name}{character, the name for the values displayed (default = "FPKM")}

\item{colorset}{character vector, the colors to interpolate between using colorRampPalette.}
}
\value{
a ggplot2 heatmap legend plot
}
\description{
Build a heatmap legend plot
}
