% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_bioconductor_dependencies}
\alias{install_bioconductor_dependencies}
\title{Retrieve dependencies from Bioconductor}
\usage{
install_bioconductor_dependencies()
}
\description{
Some packages not available (or not complete) from CRAN can be retrieved from the Bioconductor package repository:
WGCNA
}
