% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_github_deps}
\alias{install_github_deps}
\title{Install dependencies for scrattch packages from Github}
\usage{
install_github_deps()
}
\description{
Install dependencies for scrattch packages from Github
}
