% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_scrattch}
\alias{install_scrattch}
\title{Install all packages in the scrattch suite}
\usage{
install_scrattch(packages = NULL)
}
\arguments{
\item{packages}{A character vector of scrattch packages. If NULL (default), installs all packages.}
}
\description{
Install all packages in the scrattch suite
}
