% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_scrattch_deps}
\alias{install_scrattch_deps}
\title{Install all dependencies from Github and BioConductor}
\usage{
install_scrattch_deps()
}
\description{
Install all dependencies from Github and BioConductor
}
