% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildplots.R
\name{pottery_plot}
\alias{pottery_plot}
\title{Violin plots of gene expression for clusters}
\usage{
pottery_plot(genes = c("Hspa8", "Snap25", "Gad2", "Slc17a6"),
  clusters = 1:49, data_source = "internal", logscale = F, fontsize = 7,
  labelheight = 25)
}
\arguments{
\item{genes}{A character vector containing gene symbols to be plotted}

\item{clusters}{A numeric vector containing clusters to plot (for v1_anno, the range is 1:49)}

\item{data_source}{A character object defining where the data is stored. Currently only works with "internal"}

\item{logscale}{Logical object, determines if data is log scaled before plotting.}

\item{fontsize}{numeric object, the font size (in pts) used to make the plot.}

\item{labelheight}{numeric object, Percent of the plot height that should be used for the labels (0 to 100).}
}
\value{
a ggplot2 plot object
}
\description{
This function will generate plots similar to Figure 1c of Tasic, et al. (2015).
Warning: this is currently only able to work with internally-supplied datasets (v1_data and v1_anno).
Extension to user-supplied datasets will come soon.
}
\examples{
pottery_plot()

my_genes <- c("Ercc6","Ercc8","Trp53","Pgbd5")
my_clusters <- c(1,5,9,10,24,37)
pottery_plot(my_genes,my_clusters,logscale=T,fontsize=14)
}

