% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{pt2mm}
\alias{pt2mm}
\title{Convert font sizes in pt to mm}
\usage{
pt2mm(pt)
}
\arguments{
\item{pt}{A numeric font size in pt.}
}
\value{
A numeric font size in mm.
}
\description{
Convert font sizes in pt to mm
}
\examples{
pt2mm(12)

ggplot(mtcars) +
  geom_text(aes(x = mpg, y = wt, label = rownames(mtcars)),
            size = pt2mm(7))
}

