% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{riken_case}
\alias{riken_case}
\title{Convert the case of Riken genes no matter input case}
\usage{
riken_case(in_chr)
}
\arguments{
\item{in_chr}{a character vector of Riken gene symbols}
}
\value{
a character vector with correck Riken capitalization
}
\description{
Convert the case of Riken genes no matter input case
}
\examples{
test <- c("6330527o06RiK","A930038C07RIK","a330070k13rik")
riken_case(test)
}

