% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{sci_label}
\alias{sci_label}
\title{Convert integers to scientific notation labels}
\usage{
sci_label(in_num, sig_figs = 2)
}
\arguments{
\item{in_num}{a numeric vector}

\item{sig_figs}{a number indicating how many significant figures should be displayed.}
}
\value{
a character vector with numeric values reformatted in 1.2E3 format
}
\description{
Convert integers to scientific notation labels
}
\examples{
my_numbers <- c(100,15.359,32687,.000468)

sci_label(my_numbers)

sci_label(my_numbers,sig_figs=3)
}

