\name{scrattch-package}
\alias{scrattch-package}
\alias{scrattch}
\docType{package}
\title{
\packageTitle{scrattch}
}
\description{
\packageDescription{scrattch}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scrattch}
\packageIndices{scrattch}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{scrattch}

Maintainer: \packageMaintainer{scrattch}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
