% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{spiral_jitter}
\alias{spiral_jitter}
\title{Jitter x-y coordinates in a spiral pattern}
\usage{
spiral_jitter(x, y, n = NULL, max_n = NULL, radius = 1, aspect = 1,
  ratio = "golden")
}
\description{
Jitter x-y coordinates in a spiral pattern
}
