% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{split_cst}
\alias{split_cst}
\title{Split a character string by commas, spaces, and tabs into a character vector}
\usage{
split_cst(in_string)
}
\arguments{
\item{in_string}{a chr string containing commas, spaces, and/or tabs}
}
\value{
a character vector with each object separated by any combination of commas, spaces, and/or tabs
}
\description{
Split a character string by commas, spaces, and tabs into a character vector
}
\examples{
test <- "Hspa8, Scnn1a,Rbp4    Ptgs2"
split_chr(test)
}
