% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{theme_no_x}
\alias{theme_no_x}
\title{Remove the X-axis (and most other margins)}
\usage{
theme_no_x(base_size = 12, base_family = "")
}
\description{
Makes plots more suitable for use with Illustrator by removing most margins
and the X-axis (which is usually replaced by something else in my plots).
}
\details{
To further remove the space below the x-axis, use labs(x = NULL)

Based on theme_classic() from ggplot2.
}
\examples{
ggplot(mtcars) +
 geom_point(aes(x = mpg, y = wt)) +
 theme_no_x_margin() +
 labs(x = NULL)
}
