% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{title_case}
\alias{title_case}
\title{Convert the case of objects in a character vector to Title Case}
\usage{
title_case(in_chr)
}
\arguments{
\item{in_chr}{a character vector}
}
\value{
a character vector with Each Object In Title Case
}
\description{
Convert the case of objects in a character vector to Title Case
}
\examples{
test <- c("hspa8","scnn1a","fhqwghads")
title_case(test)
}

