% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v1_anno}
\alias{v1_anno}
\title{Annotations for V1 single-cell transcriptomic datasets}
\format{A Data Frame with 11 columns and 1679 rows
\describe{
 \item{filename}{A unique filename that uniquely corresponds to a single cell. Matches column names in v1_rpkm}
 \item{short_name}{A unique short cell identifier that is used in the Allen Brain Institute Vignette}
 \item{cre_id}{Numeric ID corresponding to the Cre line that the cell was harvested from}
 \item{cre_color}{Character value for a color corresponding to the cre_id}
 \item{cre_label}{Character value for a descriptive label for each cre_id}
 \item{final_id}{Numeric ID corresponding to the final cluster assignment for each cell}
 \item{final_color}{Character value for a color corresponding to final_id}
 \item{final_label}{Character value for a descriptive label for each cluster}
 \item{depth_id}{Numeric ID corresponding to the sequencing depth for each cell}
 \item{depth_color}{Character value for a color corresponding to sequencing depth}
 \item{depth_label}{Character value containing the sequencing depth in millions of reads.}
}}
\usage{
v1_anno
}
\description{
A dataset consisting of descriptive annotation data for all 1679
cells characterized in Tasic, et al. (2015).
}
\keyword{datasets}

