% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{weave_plot}
\alias{weave_plot}
\title{Generate a weave plot based on a palette of colors}
\usage{
weave_plot(palette)
}
\arguments{
\item{palette}{a character vector containing colors as either hex values (starting with #) or R colors}
}
\value{
a ggplot2 plot showing all intersections of the palette.
}
\description{
Generate a weave plot based on a palette of colors
}
