% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{get_db_data}
\alias{get_db_data}
\title{Get data from a database for scrattch plotting}
\usage{
get_db_data(data_source, genes, grouping, clusters)
}
\description{
Get data from a database for scrattch plotting
}

