% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{get_list_data}
\alias{get_list_data}
\title{Format data provided in list format for scrattch plots}
\usage{
get_list_data(datalist, genes, grouping, clusters)
}
\description{
Format data provided in list format for scrattch plots
}

