% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degenes.R
\name{get_sparse_pairwise_deg}
\alias{get_sparse_pairwise_deg}
\title{Find sparse, differentially-expressed genes in pairwise comparisons}
\usage{
get_sparse_pairwise_deg(data, anno, group_by, class1_groups, class2_groups,
  data_cut_up = 0, data_cut_dn = 0, frac_cut_up = 0.95,
  frac_cut_dn = 0.05, top = NULL, give = "genes")
}
\arguments{
\item{data}{- a data.frame containing the data to be filtered. First column must be "gene", followed by one column per sample.}

\item{anno}{- a data.frame with annotations for each sample.}

\item{group_by}{- a character object with the base of the annotation to be used for selecting groups.}

\item{class1_group}{- a numeric object}
}
\description{
Find sparse, differentially-expressed genes in pairwise comparisons
}

