% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_bioc_deps}
\alias{install_bioc_deps}
\title{Install dependencies for scrattch packages from Bioconductor}
\usage{
install_bioc_deps(packages = NULL)
}
\arguments{
\item{packages}{A character vector of scrattch packages. If NULL (default), installs dependencies for all packages.}
}
\description{
Install dependencies for scrattch packages from Bioconductor
}
