% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{values_to_colors}
\alias{values_to_colors}
\title{Convert values to colors along a color ramp}
\usage{
values_to_colors(x, minval = NULL, maxval = NULL, colorset = c("darkblue",
  "dodgerblue", "gray80", "orangered", "red"))
}
\arguments{
\item{x}{a numeric vector to be converted to colors}

\item{minval}{a number that's used to set the low end of the color scale (default = 0)}

\item{maxval}{a number that's used to set the high end of the color scale. If NULL (default), 
use the highest value in x}

\item{colorset}{a set of colors to interpolate between using colorRampPalette 
(default = c("darkblue","dodgerblue","gray80","orangered","red"))}
}
\value{
a character vector of hex color values generated by colorRampPalette. Color values will
remain in the same order as x.
}
\description{
Convert values to colors along a color ramp
}
